map function made temporarily eager
map() returning a generator is not user-friendly - requires faff with iterator_to_array and generators cannot be rewound.

This commit wraps return of map() in a vector or hash map. (If input is not a recogniseable Hopper collection, map() decides
on the return type by peeking the first key and checking if it's numeric.)

This, temporarily, fully realises the generator. Future commits will introduce lazy collections that act as generator caches.