Map and fold over HashMap call callback(, ) rather than callback([, ])
This is more compatible with PHP rather than Clojure (given PHP doesn't have function argument destructuring.)
It is easier for PHP functions to simply ignore the last argument. Furthermore, there are plenty single-argument functions
that we don't need to wrap in a destructuring anonymous function thanks to this change.

folds over \Traversable will now also follow this pattern. Incidentally, this fixes a problem in examples/book-catalog.php
which proves that [, ] was non-intuitive and non-idiomatic for PHP (if such a thing exists...)