Lazy interface introduced to indicate lazily resolved collections
map() was not user-friendly because it returned generators. I wanted a wrapper
that would act like an iterator-with-cache (sadly, CachingIterator does not work like that.)
However, I also wanted direct access to the generators due to planned performance improvements.

Solution: lazy collections now implement a Lazy interface which returns
the generator, and can therefore serve as either iterators-with-cache,
or temporary wrappers around generators.

This will require some care when handling, but as library users are not meant
to use collection methods directly, we should be safe. (...right?)